chore: move `get_active_messages/0` into PaMessages context
Creates a `Screenplay.PaMessages` context module and moves
`get_active_messages/0` from the `PaMessage` schema/model module into
that context.

Adds a new `Screenplay.PaMessages.PaMessage.Queries` module for
composable query functions of `PaMessage`s.

Uses new `PaMessage.Queries.active/3` function to build the query of
active messages.
